removed use of SystemOutRule, which wasn't working in Stress tests
